chore(ui): Layer 5b deletion sweep — verified zero-consumer primitives (Task 39)
Six cuts from audit/layer-5b-ui.md "Top 10 deletion candidates", all
verified per-export with grep across packages/, plugins/, apps/, scripts/
(excluding stories, dist, build, the file's own definition). Net −418 LOC.

DELETIONS

1. tooltip-extended.tsx (−261 LOC) — kept IconTooltip (1 consumer in
   plugins/app-companion/InferenceCloudAlertButton.tsx), deleted four
   zero-consumer exports: HoverTooltip (~103 LOC), Spotlight (~116
   LOC), TourStep + useGuidedTour (~42 LOC). Audit said the whole
   file was orphan but IconTooltip is alive — kept that piece + its
   single direct dep (Z_OVERLAY); dropped the unused Z_TOOLTIP, Button,
   and X imports that the deleted blocks pulled in.

2. skeleton.tsx (−81 LOC) — kept the Skeleton primitive (30 consumers),
   deleted SkeletonLine, SkeletonText, SkeletonCard, SkeletonChat,
   SkeletonMessage, SkeletonSidebar (zero consumers each, including
   internal-only ones). Audit said only the four "variant" skeletons
   were orphan but the strict-name grep showed Line/Text/Message/Card
   are all used only by each other.

3. status-badge.tsx StatCard (−31 LOC) — kept StatusBadge + StatusDot
   + their helpers (active), deleted StatCard. Cloud monorepo has its
   own StatCard at cloud/packages/ui — different codebase, unaffected.

4. hooks/useClickOutside.ts (−28 LOC) — only consumer was the barrel
   re-export. True orphan.

5. hooks/useKeyboardShortcuts.ts (−48 LOC) — same barrel-only pattern.
   The hook's helpers (useKeyboardShortcuts, formatShortcut,
   ShortcutConfig) were re-exported by app-core/src/hooks/useKeyboardShortcuts.ts
   and the audit was wrong that they were "zero consumers". Inlined
   the full hook body into the app-core file (which is where
   COMMON_SHORTCUTS / useShortcutsHelp already live) and dropped the
   UI file. Net: one less indirection, both files simpler, app-core
   is now self-contained on this surface.

6. ui/sonner.tsx + sonner package dep (−1 LOC + 1 dep) — the file was
   a 1-LOC re-export of sonner's Toaster; nothing in eliza mounted a
   `<Toaster />`. Cloud monorepo uses sonner extensively but in its
   own package. Removed the file, the barrel entry, and the package.json
   dependency to drop the unused install footprint.

7. tag-input.tsx (−2 LOC + 1 file) — was a 2-LOC alias re-exporting
   TagEditor under the legacy TagInput name. Migrated the single
   consumer (plugins/app-lifeops/EventEditorDrawer.tsx) to import
   TagEditor directly, deleted the alias + barrel entry.

NOT TOUCHED (audit said orphan, found alive on verification)

- form-select.tsx — audit said zero non-stories consumers; actually
  used by HeartbeatForm.tsx (app-core) and TradingPanel.tsx (examples).
- PromptDialog / usePrompt in confirm-dialog.tsx — audit said zero
  consumers; actually used by AppContext.tsx (app-core) and referenced
  in app-wallet's useWalletState.ts. The audit was outdated on this
  one — at minimum AppContext.tsx mounts <PromptDialog/> directly.
- themed-select.tsx — has the IdentitySettingsSection consumer the
  audit acknowledged; migration to Radix Select is a follow-up.
- Storybook/stories cleanup — process decision (delete all 60 vs add
  config). Out of scope for a deletion sweep.

VERIFICATION

Both `bunx tsc --noEmit` (packages/ui and packages/app-core) clean
after the changes. `bunx biome check` clean across all touched files.
Net for the PR: 12 files changed, 71 insertions, 489 deletions.
