+++ packages/ui/package.json
@@ -72,7 +72,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^1.0.0",
-    "sonner": "^2.0.7",
     "tailwind-merge": "^3.0.0"
   },
   "devDependencies": {
