+++ packages/ui/src/components/primitives/index.ts
@@ -13,13 +13,11 @@ export * from "../ui/select";
 export * from "../ui/separator";
 export * from "../ui/skeleton";
 export * from "../ui/slider";
-export * from "../ui/sonner";
 export * from "../ui/spinner";
 export * from "../ui/stack";
 export * from "../ui/switch";
 export * from "../ui/tabs";
 export * from "../ui/tag-editor";
-export * from "../ui/tag-input";
 export * from "../ui/textarea";
 export * from "../ui/tooltip";
 export * from "../ui/typography";
