--- packages/ui/src/components/ui/sonner.tsx
@@ -1,1 +1,0 @@
-export { Toaster, toast } from "sonner";
