--- packages/ui/src/components/ui/tag-input.tsx
@@ -1,2 +1,0 @@
-export type { TagEditorProps as TagInputProps } from "./tag-editor";
-export { TagEditor as TagInput } from "./tag-editor";
