+++ packages/ui/src/hooks/index.ts
@@ -1,5 +1,3 @@
-export * from "./useClickOutside";
 export * from "./useDocumentVisibility";
-export * from "./useKeyboardShortcuts";
 export * from "./useLinkedSidebarSelection";
 export * from "./useTimeout";
