--- packages/ui/src/hooks/useClickOutside.ts
@@ -1,28 +1,0 @@
-import { type RefObject, useEffect } from "react";
-
-export function useClickOutside(
-  ref: RefObject<HTMLElement | null>,
-  handler: () => void,
-  active = true,
-) {
-  useEffect(() => {
-    if (!active) return;
-
-    const onMouseDown = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) {
-        handler();
-      }
-    };
-
-    const onKeyDown = (e: KeyboardEvent) => {
-      if (e.key === "Escape") handler();
-    };
-
-    document.addEventListener("mousedown", onMouseDown);
-    document.addEventListener("keydown", onKeyDown);
-    return () => {
-      document.removeEventListener("mousedown", onMouseDown);
-      document.removeEventListener("keydown", onKeyDown);
-    };
-  }, [ref, handler, active]);
-}
