--- packages/ui/src/hooks/useKeyboardShortcuts.ts
@@ -1,48 +1,0 @@
-import { useCallback, useEffect } from "react";
-
-export interface ShortcutConfig {
-  key: string;
-  ctrl?: boolean;
-  shift?: boolean;
-  alt?: boolean;
-  meta?: boolean;
-  handler: () => void;
-  description: string;
-  scope?: string;
-}
-
-export function useKeyboardShortcuts(shortcuts: ShortcutConfig[]) {
-  const handleKeyDown = useCallback(
-    (event: KeyboardEvent) => {
-      for (const shortcut of shortcuts) {
-        const keyMatch = event.key.toLowerCase() === shortcut.key.toLowerCase();
-        const ctrlMatch = Boolean(shortcut.ctrl) === event.ctrlKey;
-        const shiftMatch = Boolean(shortcut.shift) === event.shiftKey;
-        const altMatch = Boolean(shortcut.alt) === event.altKey;
-        const metaMatch = Boolean(shortcut.meta) === event.metaKey;
-
-        if (keyMatch && ctrlMatch && shiftMatch && altMatch && metaMatch) {
-          event.preventDefault();
-          shortcut.handler();
-          break;
-        }
-      }
-    },
-    [shortcuts],
-  );
-
-  useEffect(() => {
-    window.addEventListener("keydown", handleKeyDown);
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, [handleKeyDown]);
-}
-
-export function formatShortcut(s: Omit<ShortcutConfig, "handler">): string {
-  const parts: string[] = [];
-  if (s.ctrl) parts.push("Ctrl");
-  if (s.shift) parts.push("Shift");
-  if (s.alt) parts.push("Alt");
-  if (s.meta) parts.push("Cmd");
-  parts.push(s.key.length === 1 ? s.key.toUpperCase() : s.key);
-  return parts.join("+");
-}
